Cleaned up the previous commit
