+++ Encoder/src/main/java/com/timmattison/hacking/usbrubberducky/parsers/regex/DelayInstructionParser.java
@@ -1,8 +1,8 @@
-package com.timmattison.hacking.usbrubberducky.parsers.simple;
+package com.timmattison.hacking.usbrubberducky.parsers.regex;
 
 import com.google.inject.Inject;
 import com.timmattison.hacking.usbrubberducky.constants.Whitespace;
-import com.timmattison.hacking.usbrubberducky.instructions.DefaultDelayInstruction;
+import com.timmattison.hacking.usbrubberducky.instructions.DelayInstruction;
 import com.timmattison.hacking.usbrubberducky.preprocessors.Preprocessor;
 
 import java.util.List;
@@ -14,18 +14,18 @@ import java.util.List;
  * Time: 11:32 AM
  * To change this template use File | Settings | File Templates.
  */
-public class DefaultDelayInstructionParser extends SimpleAbstractInstructionParser<DefaultDelayInstruction> {
-    private static final String name = "DEFAULT_DELAY";
+public class DelayInstructionParser extends RegexAbstractInstructionParser<DelayInstruction> {
+    private static final String name = "DELAY";
     private static final String matchingRegex = "^" + name + Whitespace.getWhitespaceCharClass() + "+" + "([0-9]+)$";
 
     @Inject
-    public DefaultDelayInstructionParser(Preprocessor preprocessor) {
+    public DelayInstructionParser(Preprocessor preprocessor) {
         super(preprocessor);
     }
 
     @Override
-    protected DefaultDelayInstruction create(List<String> input) {
-        return new DefaultDelayInstruction(Integer.parseInt(input.get(0)));
+    protected DelayInstruction create(List<String> input) {
+        return new DelayInstruction(Integer.parseInt(input.get(0)));
     }
 
     @Override
