+++ Encoder/src/test/java/com/timmattison/hacking/usbrubberducky/parsers/TestRepeatInstructionParser.java
@@ -4,7 +4,7 @@ import com.google.inject.assistedinject.Assisted;
 import com.timmattison.hacking.usbrubberducky.instructions.BasicRepeatInstruction;
 import com.timmattison.hacking.usbrubberducky.instructions.interfaces.RepeatInstruction;
 import com.timmattison.hacking.usbrubberducky.instructions.factories.RepeatInstructionFactory;
-import com.timmattison.hacking.usbrubberducky.parsers.simple.RepeatInstructionParser;
+import com.timmattison.hacking.usbrubberducky.parsers.regex.RepeatInstructionParser;
 import com.timmattison.hacking.usbrubberducky.preprocessors.LegacyPreprocessor;
 import junit.framework.Assert;
 import org.junit.Before;
